notification push and custom hook
